moved fps counter to another file containing debug imgui functions which are available without being in edit mode added editor code into an edit mode which will be later used to determine if we simulate the editor code or game code (still wip) added possiblity to show gbuffer ( still wip) added memory and compiler barriers back to the job system and removed the atomics since  this should yield better performance after I finally researched this stuff. added KeyIsPressed method to check if a key has been down and released consecutively!

+++ Dawn/Dawn/src/Dawn/Core/Config.cpp
@@ -63,7 +63,7 @@ namespace Dawn
 				ConfigFile Config = {};
 				Config.Name = File;
 
-				u32 Size = -1;
+				u64 Size = -1;
 				bool bIsParsingSection = false;
 				std::string CurrentData;
 				std::string LastSection;
@@ -167,7 +167,7 @@ namespace Dawn
 
 				// Sections and Keys are stored like this SectionName/KeyName
 				// Split these into separate parts.
-				u32 Offset = SectionAndKey.find('/');
+				u64 Offset = SectionAndKey.find('/');
 				std::string SectionPart = SectionAndKey.substr(0, Offset);
 				std::string Key = SectionAndKey.substr(Offset+1, SectionAndKey.size());
